[dnf5 plugin] config-manager: unsetvar: Log warn for unused vars

+++ dnf5-plugins/config-manager_plugin/unsetvar.cpp
@@ -64,13 +64,21 @@ void ConfigManagerUnsetVarCommand::configure() {
         }
 
         if (!std::filesystem::exists(vars_dir)) {
+            ctx.base.get_logger()->warning(
+                "config-manager: Request to remove variable but vars directory was not found: {}", vars_dir.string());
             return;
         }
 
         for (const auto & name : vars_to_remove) {
             const auto filepath = vars_dir / name;
             try {
-                std::filesystem::remove(filepath);
+                if (std::filesystem::exists(filepath)) {
+                    std::filesystem::remove(filepath);
+                } else {
+                    ctx.base.get_logger()->warning(
+                        "config-manager: Request to remove variable but it is not present in the vars directory: {}",
+                        name);
+                }
             } catch (const std::filesystem::filesystem_error & e) {
                 throw ConfigManagerError(
                     M_("Cannot remove variable file \"{}\": {}"), filepath.native(), std::string{e.what()});
